+++ CHANGELOG.md
@@ -22,7 +22,11 @@
 * Add hotlinks to the Import log summary for MARC imports (UIDATIMP-988)
 * Cover `<RecentJobLogs>` component with unit tests (UIDATIMP-720)
 * increment stripes to v7 (UIDATIMP-983)
+<<<<<<< HEAD
+* Cover `<TreeView>` component with tests (UIDATIMP-728)
+=======
 * Cover `<RecordTypesSelect>` component with unit tests (UIDATIMP-721)
+>>>>>>> 7f02f2dc92c43f806f954aaf143e90a752be610d
 
 ### Bugs fixed:
 * Auxiliary "repeatableFieldAction" property disappear while removing "vendor reference number" field mapping (UIDATIMP-987)